Added authorization in client assigned token endpoints

+++ src/endpoints/clientAssignedTokens/deleteToken/handler.ts
@@ -1,4 +1,8 @@
+import {AppResourceType, BasicCRUDActions} from '../../../definitions/system';
 import {validate} from '../../../utilities/validate';
+import {checkAuthorizatonForClientAssignedToken} from '../../contexts/authorizationChecks/checkAuthorizaton';
+import {checkOrganizationExists} from '../../organizations/utils';
+import PermissionItemQueries from '../../permissionItems/queries';
 import ClientAssignedTokenQueries from '../queries';
 import {DeleteClientAssignedTokenEndpoint} from './types';
 import {deleteClientAssignedTokenJoiSchema} from './validation';
@@ -8,9 +12,33 @@ const deleteClientAssignedToken: DeleteClientAssignedTokenEndpoint = async (
   instData
 ) => {
   const data = validate(instData.data, deleteClientAssignedTokenJoiSchema);
-  await context.session.getUser(context, instData);
+  const agent = await context.session.getAgent(context, instData);
+  const token = await context.data.clientAssignedToken.assertGetItem(
+    ClientAssignedTokenQueries.getById(data.tokenId)
+  );
+
+  const organization = await checkOrganizationExists(
+    context,
+    token.organizationId
+  );
+
+  await checkAuthorizatonForClientAssignedToken(
+    context,
+    agent,
+    organization.organizationId,
+    token,
+    BasicCRUDActions.Delete
+  );
+
   await context.data.clientAssignedToken.deleteItem(
     ClientAssignedTokenQueries.getById(data.tokenId)
+  );
+
+  await context.data.permissionItem.deleteManyItems(
+    PermissionItemQueries.getByPermissionEntity(
+      token.tokenId,
+      AppResourceType.ClientAssignedToken
+    )
   );
 };
